[Decomp][Spell] Cohesion split: extract packet senders to SpellPackets.cpp
Move the Spell:: SMSG packet-sender group (SendCastResult x2, SendSpellStart,
SendSpellGo, SendLogExecute, SendInterrupted, SendChannelUpdate/Start,
SendResurrectRequest, SendPlaySpellVisual) and their two helpers
(WriteAmmoToPacket, WriteSpellGoTargets) out of Spell.cpp into a new
SpellPackets.cpp TU. Bodies only; declarations stay in Spell.h, so call sites
and linkage are unchanged. Spell.cpp 10806 -> 9947. Pure relocation, no
behaviour change. Aligns with the ServerPacket refactor direction.
